Add custom repository implementation. Redefine all required entities.

+++ src/main/java/pl/arciemowicz/adversity/domain/AnalyticsData.java
@@ -2,12 +2,14 @@ package pl.arciemowicz.adversity.domain;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class AnalyticsData {
 
     private String dataSource;
     private String campaign;
     private String date;
-    private long clicks;
-    private long impressions;
+    private BigDecimal clicks;
+    private BigDecimal impressions;
 }
