+++ src/main/java/pl/arciemowicz/adversity/domain/dto/Impression.java
@@ -2,11 +2,12 @@ package pl.arciemowicz.adversity.domain.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
 public class Impression {
 
     private LocalDate date;
-    private long impressions;
+    private BigDecimal impressions;
 }
